MINIFICPP-601: Add Convert Scheduling structure to millis
